refactor(library/Parser): renames "Parser/StaticParser.ts" to "Parser/Static.ts"
