+++ src/library/Parser/Static.ts
