+++ src/library/Parser/string/StaticStringParser.ts
@@ -2,7 +2,7 @@ import type ParsingError from '@/library/ParsingError';
 
 import type {
   Parser_Static,
-} from '../StaticParser';
+} from '../Static';
 
 type Parser_StaticString<
   SomeParsedOutput,
